requirements.txt
discord.py==2.3.2
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0